requirements.txt
dj-database-url==0.3.0
Django==1.8.4
django-extensions==1.5.6
django-filter==0.11.0
djangorestframework==3.2.3
djangorestframework-jwt==1.7.2
gunicorn==19.3.0
psycopg2==2.6.1
py==1.4.30
PyJWT==1.4.0
pytest==2.7.2
pytest-django==2.8.0
six==1.9.0
wheel==0.24.0
